feat(general): Finished project for made RTP Port

+++ src/utils/main_utils.c
@@ -6,16 +6,16 @@
 
 #include <math.h>
 
-void generate_sin_components(time_t timestamp, float *day_sin, float *minute_sin) {
+void generate_sin_components(time_t timestamp, double *day_sin, double *minute_sin) {
     struct tm *tm = gmtime(&timestamp);
     *day_sin = sin((2 * M_PI * tm->tm_yday) / 365.0);
     *minute_sin = sin((2 * M_PI * (tm->tm_hour * 60 + tm->tm_min)) / 1440.0);
 }
 
-float normalize(float value, float min, float max) {
+double normalize(double value, double min, double max) {
     return (value - min) / (max - min);
 }
 
-float denormalize(float value, float min, float max) {
+double denormalize(double value, double min, double max) {
     return value * (max - min) + min;
 }
